[Fix][Flink-20] Dedup starter deps and restore sink2 metrics
